Example of automated a11y testing in CI This is a proof-of-concept for integrating automated a11y scanning into our CI pipeline.
It creates a new Circle Job called accessibility which will run in parallel to the test and build jobs.

The job uses a dotnet docker image because it needs to run a local version of a dotnet app for pa11y to test.

node/npm and supporting packages for puppeteer (which powers the headless browsing for pa11y) are installed. Then it runs the app, waits for the app to spin up, then runs pa11y-ci on localhost:5000.

Currently this job only runs pa11y for the Dashboard app.