Fix: move GSA IP restrictions from web app level to Front Door/WAF policy level Initially, restricting access to only GSA IP's was set on the web apps themselves.
When Azure Front Doors were applied to the apps, this blocked traffic from front door backends.

Instead, the web apps are now restricted to AzureFrontDoor.Backend traffic, and GSA IP restrictions are set as rules on the related Front Door WAF policies.

This creates a co-dependency when creating web app and front door resources: web apps need the front door ID in order to set the backend restrictions, and front doors need the app host name in order to connect the app to the front door resource.

Consequently, the ARM templates become more reliant on naming conventions. It also becomes more difficult to name resources dynamically through the templates, as names are often needed for other resources before the app resource is deployed.