Fix call to dotnet list package
`dotnet list package` will throw a System.ArgumentException if it is
passed an empty string as an argument. To address this and avoid a
shellcheck conformance flag, use arrays to build up the option string.
But implement it so that there is always at least one value in the
option array to avoid this issue[1] in versions of bash < 4.4.

Apply a similar technique to other places in the code for consistency
and to just avoid passing empty arguments where none should be present.

[1] https://stackoverflow.com/a/61551944