Add read-only role to participants DB cluster
- Create NOLOGIN role with permission to connect to each state database
- Grant role SELECT permission to necessary tables

Intent is for this role to be granted to the forthcoming orchestrator
managed identity role (and potentially a future bulk match identity
role).

Creating this role/group may be overkill if it's used by only one or two
identities, but this approach avoids making `create-databases.bash`
dependent on other resources.