Add delay after SIEM app regstration creation
Our standard 60 second delay *should* allow for the SIEM app
registration to be valid by the time it is passed to the event hub
ARM template. Otherwise, we will get an error code of InvalidPrincipalId
and a detailed error message of:
"A valid principal ID must be provided for role assignment."
in the deployment error log for the rg-core-<ENV> resource group.